+++ crossdata-cassandra/src/main/scala/com/stratio/crossdata/sql/sources/cassandra/CassandraQueryProcessor.scala
@@ -28,8 +28,7 @@ import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.planning.PhysicalOperation
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.sources
-import org.apache.spark.sql.sources.crossdata.CatalystToCrossdataAdapter
-import org.apache.spark.sql.sources.{Filter => SourceFilter}
+import org.apache.spark.sql.sources.{Filter => SourceFilter, CatalystToCrossdataAdapter}
 import org.apache.spark.sql.types.UTF8String
